requirements.txt
# AI Product Manager Environment - Requirements
# Updated for Windows compatibility (pre-built wheels)

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# API Client & Async
httpx>=0.26.0
aiohttp>=3.9.2

# LLM Integrations
groq>=0.4.0
openai>=1.12.0

# Environment & Configuration
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Utilities
typing-extensions>=4.9.0

# Development (optional)
pytest>=7.4.3
pytest-asyncio>=0.23.0
black>=24.1.0
flake8>=7.0.0
